3.1-drf-intro/smart_home/requirements.txt
Pillow==9.5.0
django
djangorestframework
psycopg2-binary
